@@ -13,122 +13,122 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.0.7",
+    "@angular/animations": "^9.1.2",
     "@angular/cdk": "^9.1.3",
-    "@angular/common": "^9.0.7",
-    "@angular/compiler": "^9.0.7",
-    "@angular/core": "^9.0.7",
+    "@angular/common": "^9.1.2",
+    "@angular/compiler": "^9.1.2",
+    "@angular/core": "^9.1.2",
     "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "^9.0.7",
-    "@angular/material": "^9.1.3",
-    "@angular/platform-browser": "^9.0.7",
-    "@angular/platform-browser-dynamic": "^9.0.7",
-    "@angular/router": "^9.0.7",
+    "@angular/forms": "^9.1.2",
+    "@angular/material": "^9.2.1",
+    "@angular/platform-browser": "^9.1.2",
+    "@angular/platform-browser-dynamic": "^9.1.2",
+    "@angular/router": "^9.1.2",
     "@auth0/angular-jwt": "^4.0.0",
-    "@date-io/date-fns": "^2.4.0",
+    "@date-io/date-fns": "^2.6.1",
     "@flowjs/flow.js": "^2.14.0",
     "@flowjs/ngx-flow": "^0.4.3",
-    "@mat-datetimepicker/core": "^4.0.1",
-    "@material-ui/core": "^4.9.3",
+    "@mat-datetimepicker/core": "^4.1.0",
+    "@material-ui/core": "^4.9.11",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/pickers": "^3.2.10",
-    "@ngrx/effects": "^8.6.0",
-    "@ngrx/store": "^8.6.0",
-    "@ngrx/store-devtools": "^8.6.0",
+    "@ngrx/effects": "^9.1.0",
+    "@ngrx/store": "^9.1.0",
+    "@ngrx/store-devtools": "^9.1.0",
     "@ngx-share/core": "^7.1.4",
-    "@ngx-translate/core": "^12.1.1",
+    "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^4.0.0",
-    "ace-builds": "^1.4.8",
-    "angular-gridster2": "^9.0.1",
-    "angular2-hotkeys": "^2.1.5",
+    "ace-builds": "^1.4.11",
+    "angular-gridster2": "^9.1.0",
+    "angular2-hotkeys": "^2.2.0",
     "base64-js": "^1.3.1",
-    "canvas-gauges": "^2.1.5",
+    "canvas-gauges": "^2.1.7",
     "compass-sass-mixins": "^0.12.7",
-    "core-js": "^3.6.4",
-    "date-fns": "^2.9.0",
+    "core-js": "^3.6.5",
+    "date-fns": "^2.12.0",
     "flot": "git://github.com/thingsboard/flot.git#0.9-work",
     "flot.curvedlines": "git://github.com/MichaelZinsmaier/CurvedLines.git#master",
     "font-awesome": "^4.7.0",
     "javascript-detect-element-resize": "^0.5.3",
-    "jquery": "^3.4.1",
-    "jquery.terminal": "^2.14.0",
-    "js-beautify": "^1.10.3",
+    "jquery": "^3.5.0",
+    "jquery.terminal": "^2.15.4",
+    "js-beautify": "^1.11.0",
     "json-schema-defaults": "^0.4.0",
     "jstree": "^3.3.9",
     "jstree-bootstrap-theme": "^1.0.1",
-    "jszip": "^3.2.2",
+    "jszip": "^3.4.0",
     "leaflet": "^1.6.0",
     "leaflet-geometryutil": "^0.9.3",
     "leaflet-polylinedecorator": "^1.6.0",
     "leaflet-providers": "^1.9.1",
-    "leaflet.gridlayer.googlemutant": "^0.8.0",
+    "leaflet.gridlayer.googlemutant": "^0.9.0",
     "leaflet.markercluster": "^1.4.1",
     "material-design-icons": "^3.0.1",
     "messageformat": "^2.3.0",
     "moment": "^2.24.0",
-    "ngx-clipboard": "^12.3.1",
-    "ngx-color-picker": "^9.0.0",
-    "ngx-daterangepicker-material": "^2.2.0",
+    "ngx-clipboard": "^13.0.0",
+    "ngx-color-picker": "^9.1.0",
+    "ngx-daterangepicker-material": "^3.0.1",
     "ngx-flowchart": "git://github.com/thingsboard/ngx-flowchart.git#master",
     "ngx-hm-carousel": "^2.0.0-rc.1",
-    "ngx-translate-messageformat-compiler": "^4.5.0",
+    "ngx-translate-messageformat-compiler": "^4.6.0",
     "objectpath": "^2.0.0",
     "prop-types": "^15.7.2",
     "raphael": "^2.3.0",
-    "rc-select": "^10.0.0-rc.6",
-    "react": "^16.12.0",
-    "react-ace": "^8.0.0",
-    "react-dom": "^16.12.0",
-    "react-dropzone": "^10.2.1",
+    "rc-select": "^10.2.4",
+    "react": "^16.13.1",
+    "react-ace": "^8.1.0",
+    "react-dom": "^16.13.1",
+    "react-dropzone": "^10.2.2",
     "reactcss": "^1.2.3",
-    "rxjs": "^6.5.4",
+    "rxjs": "^6.5.5",
     "schema-inspector": "1.6.8",
     "screenfull": "^5.0.2",
     "split.js": "^1.5.11",
     "systemjs": "0.21.5",
     "tinycolor2": "^1.4.1",
     "tooltipster": "^4.2.7",
-    "tslib": "^1.10.0",
+    "tslib": "^1.11.1",
     "tv4": "^1.3.0",
     "typeface-roboto": "^0.0.75",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^9.0.0-beta.7",
-    "@angular-devkit/build-angular": "^0.900.7",
-    "@angular/cli": "^9.0.7",
-    "@angular/compiler-cli": "^9.0.7",
-    "@angular/language-service": "^9.0.7",
+    "@angular-builders/custom-webpack": "^9.1.0",
+    "@angular-devkit/build-angular": "^0.901.1",
+    "@angular/cli": "^9.1.1",
+    "@angular/compiler-cli": "^9.1.2",
+    "@angular/language-service": "^9.1.2",
     "@types/canvas-gauges": "^2.1.2",
     "@types/flot": "^0.0.31",
-    "@types/jasmine": "^3.5.5",
+    "@types/jasmine": "^3.5.10",
     "@types/jasminewd2": "^2.0.8",
-    "@types/jquery": "^3.3.32",
-    "@types/js-beautify": "^1.8.1",
+    "@types/jquery": "^3.3.35",
+    "@types/js-beautify": "^1.8.2",
     "@types/jstree": "^3.3.39",
     "@types/jszip": "^3.1.7",
-    "@types/leaflet": "^1.5.9",
+    "@types/leaflet": "^1.5.12",
     "@types/leaflet-polylinedecorator": "^1.6.0",
-    "@types/lodash": "^4.14.149",
-    "@types/raphael": "^2.1.30",
-    "@types/react": "^16.9.20",
-    "@types/react-dom": "^16.9.5",
+    "@types/lodash": "^4.14.150",
+    "@types/raphael": "^2.3.0",
+    "@types/react": "^16.9.34",
+    "@types/react-dom": "^16.9.6",
     "@types/tinycolor2": "^1.4.2",
     "@types/tooltipster": "^0.0.29",
-    "codelyzer": "^5.2.1",
+    "codelyzer": "^5.2.2",
     "compression-webpack-plugin": "^3.1.0",
     "directory-tree": "^2.2.4",
     "jasmine-core": "^3.5.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.4.1",
+    "jasmine-spec-reporter": "^5.0.1",
+    "karma": "^5.0.2",
     "karma-chrome-launcher": "^3.1.0",
     "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "^3.1.1",
-    "karma-jasmine-html-reporter": "^1.5.2",
+    "karma-jasmine-html-reporter": "^1.5.3",
     "ngrx-store-freeze": "^0.2.4",
-    "protractor": "^5.4.3",
-    "ts-node": "^8.6.2",
-    "tslint": "^6.0.0",
+    "protractor": "^5.4.4",
+    "ts-node": "^8.9.0",
+    "tslint": "^6.1.1",
     "typescript": "~3.7.5"
   }
 }
