@@ -50,7 +50,7 @@ export class HomeComponent extends PageComponent implements AfterViewInit, OnIni
   sidenavMode: 'over' | 'push' | 'side' = 'side';
   sidenavOpened = true;
 
-  logo = require('../../../assets/logo_title_white.svg');
+  logo = require('../../../assets/logo_title_white.svg').default;
 
   @ViewChild('sidenav')
   sidenav: MatSidenav;
