@@ -31,7 +31,6 @@ import {
 import { Router } from '@angular/router';
 import { DialogComponent } from '@app/shared/components/dialog.component';
 import { DashboardState } from '@app/shared/models/dashboard.models';
-import { MatDialog } from '@angular/material/dialog';
 import { DashboardStateInfo } from '@home/pages/dashboard/states/manage-dashboard-states-dialog.component.models';
 import { TranslateService } from '@ngx-translate/core';
 import { DashboardUtilsService } from '@core/services/dashboard-utils.service';
@@ -71,8 +70,7 @@ export class DashboardStateDialogComponent extends
               public dialogRef: MatDialogRef<DashboardStateDialogComponent, DashboardStateInfo>,
               private fb: FormBuilder,
               private translate: TranslateService,
-              private dashboardUtils: DashboardUtilsService,
-              private dialog: MatDialog) {
+              private dashboardUtils: DashboardUtilsService) {
     super(store, router, dialogRef);
 
     this.states = this.data.states;
